rest-api-with-DRF/requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==23.2
djangorestframework==3.16.0
pytz==2023.3
sqlparse==0.4.4
